Upload of Sinoatrial Node Model implementation
This commit includes the initial upload of my Sinoatrial Node (SAN) Model implementation in Python. The model is based on the work by Fabbri et al.  [1] and provides a computational analysis of the human sinus node action potential. The code is organized into a structured directory with separate modules for the model, helper functions, and an example notebook. The 'config.json' file contains the model parameters.

Feel free to explore, use, and contribute to this project. Proper attribution to the original authors and the Fabbri et al. paper [1] is appreciated.

[Optional: Add any relevant tags or references here]

References:
1. Fabbri, Alan, et al. "Computational analysis of the human sinus node action potential: model development and effects of mutations." The Journal of Physiology 595.7 (2017): 2365-2396. DOI: 10.1113/JP273259